feat: Add user input and grid alignment selection
Add user input functionality to select a starting grid alignment. Used
pattern matching to ensure input is properly cleaned and exceptions are
handled. Refactor code to separate game and grid helper functions in a new
module. Import necessary items to the `main.rs` file.
